I changed some things but I forgot what I changed :)

+++ src/main/java/SpaceLayer.java
@@ -5,7 +5,7 @@
  */
 public class SpaceLayer extends Layer {
 
-    int dy;
+    float dy;
 
     public SpaceLayer(){
         this.dy = 2;
